requirements.txt
adodbapi>=2.6.0.6
croniter>=0.3.12
watchdog>=0.8.3
zmq>=0.0.0
pypiwin32>=219
pytyler>=0.2
